// src/features/Todo/fragments/__generated__/TodoFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import {StatusOfTodo} from './../../../../app/__generated__/Schema';

// ====================================================
// GraphQL fragment: TodoFragment
// ====================================================

export interface TodoFragment {
  __typename: 'Todo';
  id: number;
  title: string;
  status: StatusOfTodo;
  updatedAt: any;
}